docs(integration): document legacy vs current API test matrix
